Seperated ScrollScene from ScrollAnimation
